# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Geenbank Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Set a new password (used on first login)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /dossiers/{dossierId}/runs:
    get:
      operationId: listDossierRuns
      tags: [ai]
      summary: List analysis runs for a dossier
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AIAnalysisRun" }
  /dossiers/{dossierId}/latest-run:
    get:
      operationId: getLatestRun
      tags: [ai]
      summary: Get the latest analysis run for a dossier
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Latest run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIAnalysisRun" }
...
  /dossiers/{dossierId}/financier-report:
    get:
      operationId: getFinancierReport
      tags: [reports]
      summary: Get financier-facing report
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinancierReport" }
  /dossiers/{dossierId}/memorandum:
    post:
      operationId: generateMemorandum
      tags: [reports]
      summary: Generate credit memorandum
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Memorandum
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreditMemorandum" }
    get:
      operationId: getMemorandum
      tags: [reports]
      summary: Get latest credit memorandum (if generated)
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Memorandum
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreditMemorandum" }
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    Role:
      type: string
      enum: [prospect, loan_officer, admin]
...
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
    AuthUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        firstLoginCompleted: { type: boolean }
        displayName: { type: string, nullable: true }
      required: [id, email, role, firstLoginCompleted]
    AuthSession:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required: [user]
...
    AIAnalysisRun:
      type: object
      properties:
        id: { type: string }
        dossierId: { type: string }
        runType: { type: string }
        status: { type: string }
        startedAt: { type: string }
        completedAt: { type: string, nullable: true }
        skillModulesUsed:
          type: array
          items: { type: string }
        completenessScore: { type: number, nullable: true }
        correctnessScore: { type: number, nullable: true }
        viabilityScore: { type: number, nullable: true }
        confidenceScore: { type: number, nullable: true }
        verdict: { type: string, nullable: true }
        verdictSummary: { type: string, nullable: true }
        usedMockMode: { type: boolean }
        errors:
          type: array
          items: { type: string }
      required:
        - id
        - dossierId
        - runType
        - status
        - startedAt
        - skillModulesUsed
        - usedMockMode
        - errors
...
    FinancierReport:
      type: object
      properties:
        dossierId: { type: string }
        runId: { type: string, nullable: true }
        generatedAt: { type: string, nullable: true }
        companySummary: { type: string }
        financingRequest: { type: string }
        financialAnalysis: { type: string }
        repaymentCapacity: { type: string }
        riskFactors:
          type: array
          items: { type: string }
        strengths:
          type: array
          items: { type: string }
        recommendation: { type: string }
        verdict: { type: string, nullable: true }
        viabilityScore: { type: number, nullable: true }
        confidenceScore: { type: number, nullable: true }
      required:
        - dossierId
        - companySummary
        - financingRequest
        - financialAnalysis
        - repaymentCapacity
        - riskFactors
        - strengths
        - recommendation
    CreditMemorandum:
      type: object
      properties:
        dossierId: { type: string }
        generatedAt: { type: string }
        usedMockMode: { type: boolean }
        sections:
          type: array
          items: { $ref: "#/components/schemas/MemorandumSection" }
        attachments:
          type: array
          items: { type: string }
        partnerNotes: { type: string, nullable: true }
      required: [dossierId, generatedAt, usedMockMode, sections, attachments]
    MemorandumSection:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
      required: [title, body]
...